added check user on token
